fix(openid4vc-client): set package to private (#1210)
fix: fix typing issues with typescript 4.9 (#1214)
Fixes https://github.com/hyperledger/aries-framework-javascript/issues/1205
resolver (#1247)
fix: set updateAt on records when updating a record (#1272)
fix(transport)!: added docs moved connection to connectionId (#1222)
fix(indy-vdr): export relevant packages from root (#1291)
fix(askar): generate nonce suitable for anoncreds (#1295)
resolver and registrar for did:indy (#1253)
fix: imports from core (#1303)
fix: thread id improvements (#1311)
fix: loosen base64 validation (#1312)
fix(samples): dummy module response message type (#1321)
fix: seed and private key validation and return type in registrars (#1324)
fix!: don't emit legacy did:sov prefix for new protocols (#1245)
fix(askar): anoncrypt messages unpacking (#1332)
fix: expose indy pool configs and action menu messages (#1333)
fix: create new socket if socket state is 'closing' (#1337)
fix(anoncreds): include prover_did for legacy indy (#1342)
fix(indy-sdk): import from core (#1346)
fix(anoncreds-rs): save revocation registry index (#1351)
fix: isNewSocket logic (#1355)
fix(tenant): Correctly configure storage for multi tenant agents (#1359)
Fixes hyperledger#1353
fix(anoncreds): Buffer not imported from core (#1367)
fix(core): repository event when calling deleteById (#1356)
fix(askar): custom error handling (#1372)
fix: return HTTP 415 if unsupported content type (#1313)
fix: remove named capture groups (#1378)
fix example usage of indy-sdk-react-native package (#1382)
fix: connection id in sessions for new connections (#1383)
fix: did cache key not being set correctly (#1394)
fix: incorrect type for anoncreds registration (#1396)
fix: reference to indyLedgers in IndyXXXNotConfiguredError (#1397)
fix: add reflect-metadata (#1409)
fix: various anoncreds revocation fixes (#1416)
fix: jsonld credential format identifier version (#1412)
fix: remove `deleteOnFinish` and added documentation (#1418)
fix(askar): default key derivation method (#1420)
fix(anoncreds): make revocation status list inline with the spec (#1421)
fix(anoncreds-rs): revocation status list as JSON (#1422)
fix: issuance with unqualified identifiers (#1431)
fix(connections): store imageUrl when using DIDExchange (#1433)
fix(indy-vdr): do not force indy-vdr version (#1434)
fix: small issues with migration and WAL files (#1443)
fix: migration of link secret (#1444)
fix: Emit RoutingCreated event for mediator routing record (#1445)
fix: small updates to cheqd module and demo (#1439)
fix: remove scope check from response (#1450)

+++ packages/node/src/NodeFileSystem.ts
@@ -1,23 +1,34 @@
-import type { FileSystem } from '@aries-framework/core'
+import type { DownloadToFileOptions, FileSystem } from '@aries-framework/core'
 
+import { AriesFrameworkError, TypedArrayEncoder } from '@aries-framework/core'
+import { createHash } from 'crypto'
 import fs, { promises } from 'fs'
 import http from 'http'
 import https from 'https'
-import { tmpdir } from 'os'
+import { tmpdir, homedir } from 'os'
 import { dirname } from 'path'
 
-const { access, readFile, writeFile } = promises
+const { access, readFile, writeFile, mkdir, rm, unlink, copyFile } = promises
 
 export class NodeFileSystem implements FileSystem {
-  public readonly basePath
+  public readonly dataPath
+  public readonly cachePath
+  public readonly tempPath
 
   /**
    * Create new NodeFileSystem class instance.
    *
-   * @param basePath The base path to use for reading and writing files. process.cwd() if not specified
+   * @param baseDataPath The base path to use for reading and writing data files used within the framework.
+   * Files will be created under baseDataPath/.afj directory. If not specified, it will be set to homedir()
+   * @param baseCachePath The base path to use for reading and writing cache files used within the framework.
+   * Files will be created under baseCachePath/.afj directory. If not specified, it will be set to homedir()
+   * @param baseTempPath The base path to use for reading and writing temporary files within the framework.
+   * Files will be created under baseTempPath/.afj directory. If not specified, it will be set to tmpdir()
    */
-  public constructor(basePath?: string) {
-    this.basePath = basePath ?? tmpdir()
+  public constructor(options?: { baseDataPath?: string; baseCachePath?: string; baseTempPath?: string }) {
+    this.dataPath = options?.baseDataPath ? `${options?.baseDataPath}/.afj` : `${homedir()}/.afj/data`
+    this.cachePath = options?.baseCachePath ? `${options?.baseCachePath}/.afj` : `${homedir()}/.afj/cache`
+    this.tempPath = `${options?.baseTempPath ?? tmpdir()}/.afj`
   }
 
   public async exists(path: string) {
@@ -29,9 +40,17 @@ export class NodeFileSystem implements FileSystem {
     }
   }
 
+  public async createDirectory(path: string): Promise<void> {
+    await mkdir(dirname(path), { recursive: true })
+  }
+
+  public async copyFile(sourcePath: string, destinationPath: string): Promise<void> {
+    await copyFile(sourcePath, destinationPath)
+  }
+
   public async write(path: string, data: string): Promise<void> {
     // Make sure parent directories exist
-    await promises.mkdir(dirname(path), { recursive: true })
+    await mkdir(dirname(path), { recursive: true })
 
     return writeFile(path, data, { encoding: 'utf-8' })
   }
@@ -40,13 +59,18 @@ export class NodeFileSystem implements FileSystem {
     return readFile(path, { encoding: 'utf-8' })
   }
 
-  public async downloadToFile(url: string, path: string) {
+  public async delete(path: string): Promise<void> {
+    await rm(path, { recursive: true, force: true })
+  }
+
+  public async downloadToFile(url: string, path: string, options: DownloadToFileOptions) {
     const httpMethod = url.startsWith('https') ? https : http
 
     // Make sure parent directories exist
-    await promises.mkdir(dirname(path), { recursive: true })
+    await mkdir(dirname(path), { recursive: true })
 
     const file = fs.createWriteStream(path)
+    const hash = options.verifyHash ? createHash('sha256') : undefined
 
     return new Promise<void>((resolve, reject) => {
       httpMethod
@@ -56,15 +80,32 @@ export class NodeFileSystem implements FileSystem {
             reject(`Unable to download file from url: ${url}. Response status was ${response.statusCode}`)
           }
 
+          hash && response.pipe(hash)
           response.pipe(file)
-          file.on('finish', () => {
+          file.on('finish', async () => {
             file.close()
+
+            if (hash && options.verifyHash?.hash) {
+              hash.end()
+              const digest = hash.digest()
+              if (digest.compare(options.verifyHash.hash) !== 0) {
+                await fs.promises.unlink(path)
+
+                reject(
+                  new AriesFrameworkError(
+                    `Hash of downloaded file does not match expected hash. Expected: ${
+                      options.verifyHash.hash
+                    }, Actual: ${TypedArrayEncoder.toUtf8String(digest)})}`
+                  )
+                )
+              }
+            }
             resolve()
           })
         })
         .on('error', async (error) => {
           // Handle errors
-          await fs.promises.unlink(path) // Delete the file async. (But we don't check the result)
+          await unlink(path) // Delete the file async. (But we don't check the result)
           reject(`Unable to download file from url: ${url}. ${error.message}`)
         })
     })
